Lesson 1 task 4 fly Gagarin 2

+++ src/main/kotlin/lesson1_task4.kt
@@ -4,8 +4,8 @@ fun main () {
     val length: Long = 40868600000
     val age: Byte = 27
     val partOfday: Float = 0.075f
-    val seconds: Int = 6480
-    val partOfYear: Float = 2.0547945205479453E-4f
+    val seconds: Short = 6480
+    val partOfYear: Double = 2.0547945205479453E-4
     val apogee: Int = 327000
 
     println(length)
